Wrapper based missing data handling (untested)
MissingEmissionFunction class wraps other EmissionFunction subclasses with a check for missing data. Meant to be used together with IterMissing subclass of Iter.

IterMissing uses pointer arithmetic to get at appropriate missing data slot. Decision to not make "next()/prev()" virtual was motivated by idea of not making the base case worst. However, it may be that the cost of a virtual call to next()/prev() is a reasonable price to pay to avoid the more costly implementation of is_missing().