DiscreteGamma optimization (pre-compute the first 64 values)
Assuming that (x + _offset) is an integer >= 0; then we can pre-compute the first K(=64) values of the log-probability to speed up inference.

Table size (K above) is settable as an option (tblSize).
NOTE: It must be set to the same value for all instances in the same emission group!

This optimization assumes most values are small. Adjust tblSize as needed.